requirements.txt
anthropic==0.31.2
boto3==1.34.148
folium==0.17.0
geopy==2.4.1
googlemaps==4.10.0
hugchat==0.4.10
Pillow==9.5.0
Pillow==10.4.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
streamlit==1.36.0
streamlit_folium==0.21.1
